Refactor away the API module.
