--- src/api/tests/mod.rs
@@ -1,1 +1,0 @@
-mod line_draw_fn;
